Refactored project to have proper file structure, added SBT build support, and refactored Parser.java to use NIO where possible.
