+++ build.sbt
@@ -1,0 +1,7 @@
+name := "quiz"
+
+organization := "io.teamed"
+
+version := "1.0.0-SNAPSHOT"
+
+sbtVersion := "0.13.11"
